+++ federatedscope/core/data/__init__.py
@@ -1,4 +1,8 @@
 from federatedscope.core.data.base_data import StandaloneDataDict, ClientData
-from federatedscope.core.data.base_translator import BaseDataTranslator
+from federatedscope.core.data.base_translator import BaseDataTranslator, \
+    DummyDataTranslator
 
-__all__ = ['StandaloneDataDict', 'ClientData', 'BaseDataTranslator']
+__all__ = [
+    'StandaloneDataDict', 'ClientData', 'BaseDataTranslator',
+    'DummyDataTranslator'
+]
